Add Data layer
This commit adds the Data Layer of the application. PostgreSQL
is used for the db management. Postgres has been successfully downloaded
and installed. JDBC driver for postgres has been downloaded and installed.
Installed database perspective for eclipse. Utilizing db perspective,
the driver and connection details, a successful connection to the db
was established.