Add POST functionality and save to db
This commit adds the POST request functionality to our application.
This allows the addition of students to the db. There is a small
validation process that checks if the email is already in the db.
If the student already exists, it will throw an exception with a 500
error, along with a message letting the client know the email is taken.

the application properties file was adjusted to allow for our server
to include messages.