Add PUT request functionality - Update student name and email
This commit adds the update student endpoint.  Users can now update
the students name and email.
Simple validation has been added to the student service layer for
business logic needs. 1.) check for valid student id. 2.) make sure
the user has not entered blank credentials. 3.) check to make sure
the user isn't trying to change the name or email to what is currently is.
Once the information has passed validation the setter for the Student entity
will be used to update the name/email in the database.